+++ backend/config.yaml
@@ -4,5 +4,3 @@ elem_despesa_path: 'data/elem_despesa.txt'
 parquet_path: 'data/tce.parquet'
 
 embedding_model: 'sentence-transformers/all-MiniLM-L12-v1'
-
-itens_por_pagina: 10
