+++ frontend/react/src/elemdespesas.json
@@ -1,0 +1,131 @@
+[
+  "OUTROS SERVICOS DE TERCEIROS   PESSOA JURIDICA",
+  "INDENIZACOES E RESTITUICOES TRABALHISTAS   ATIVOS",
+  "DIARIAS   CIVIL",
+  "DESP DE EXERCICIOS ANTERIORES RELATIVAS A PESSOAL E ENCARGOS DE SERVIDORES ATIVOS",
+  "SENTENCAS JUDICIAIS",
+  "MATERIAL DE CONSUMO",
+  "INDENIZACOES E RESTITUICOES",
+  "CONTRIBUICAO PARA O FGTS",
+  "CONTRIBUICAO PARA O REGIME GERAL DE PREVIDENCIA  INSS ",
+  "DESPESAS DE EXERCICIOS ANTERIORES",
+  "OBRIGACOES TRIBUTARIAS E CONTRIBUTIVAS",
+  "PRINCIPAL DA DIVIDA CONTRATUAL RESGATADO",
+  "JUROS SOBRE A DIVIDA POR CONTRATO",
+  "OUTROS ENCARGOS SOBRE A DIVIDA POR CONTRATO",
+  "OUTROS SERVICOS DE TERCEIROS   PESSOA FISICA",
+  "SERVICOS DE TECNOLOGIA DA INFORMACAO E COMUNICACAO   PESSOA JURIDICA",
+  "OBRAS E INSTALACOES",
+  "EQUIPAMENTOS E MATERIAL PERMANENTE",
+  "PASSAGENS E DESPESAS COM LOCOMOCAO",
+  "TRANSFERENCIAS A INSTITUICOES PRIVADAS SEM FINS LUCRATIVOS",
+  "MATERIAL  BEM OU SERVICO PARA DISTRIBUICAO GRATUITA",
+  "SERVICOS DE CONSULTORIA",
+  "OUTROS ENCARGOS PATRONAIS",
+  "CONTRIBUICOES PREVIDENCIARIAS   RPPS   PESSOAL ATIVO   PLANO PREVIDENCIARIO",
+  "RESSARCIMENTO DE DESPESAS DE PESSOAL REQUISITADO",
+  "AUXILIO ALIMENTACAO",
+  "EXECUCAO DE CONTRATO DE PPP",
+  "OUTRAS DESPESAS DE EXERCICIOS ANTERIORES",
+  "OUTROS AUXILIOS FINANCEIROS A PESSOAS FISICAS",
+  "CONTRIBUICAO PARA O REGIME PROPRIO DE PREVIDENCIA  RPPS ",
+  "CONTRIBUICOES PREVIDENCIARIAS   RPPS   PESSOAL ATIVO   PLANO FINANCEIRO",
+  "PREMIACOES CULTURAIS  ARTISTICAS  CIENTIFICAS  DESPORTIVAS E OUTRAS",
+  "VENCIMENTOS E VANTAGENS FIXAS   PESSOAL CIVIL",
+  "PAGAS COM RECURSOS DO TESOURO MUNICIPAL",
+  "CONTRATACAO POR TEMPO DETERMINADO",
+  "OUTROS BENEFICIOS PREVIDENCIARIOS DO SERVIDOR OU DO MILITAR",
+  "SALARIO FAMILIA",
+  "VENCIMENTOS E VANTAGENS FIXAS  EXCETO SALARIO MATERNIDADE E LICENCA SAUDE  PESSOAL CIVIL",
+  "SALARIO MATERNIDADE",
+  "LICENCA SAUDE",
+  "OUTROS OBRIGACOES PATRONAIS",
+  "DESP DE EXERCICIOS ANTERIORES RELATIVAS A PESSOAL E ENCARGOS DE INATIVOS",
+  "CONTRIBUICAO PARA O REGIME PROPRIO DE PREVIDENCIA  RPPS    PESSOAL ATIVO",
+  "EXECUCAO ORCAMENTARIA DELEGADA A ESTADOS E AO DISTRITO FEDERAL",
+  "VENCIMENTOS E VANTAGENS FIXAS PESSOAL CIVIL",
+  "OUTRAS DESPESAS VARIAVEIS   PESSOAL CIVIL",
+  "ELEMENTO GENERICO   TRANSF  UNIAO",
+  "SENTENCAS JUDICIAIS RELATIVAS A PESSOAL E ENCARGOS DE SERVIDORES ATIVOS",
+  "SUBVENCOES ECONOMICAS",
+  "AQUISICAO DE IMOVEIS",
+  "PENSOES DO RPPS E DO MILITAR",
+  "APOSENTADORIAS DO RPPS  RESERVA REMUNERADA E REFORMAS DOS MILITARES",
+  "PAGOS COM RECURSOS VINCULADOS RPPS",
+  "CONTRIBUICOES",
+  "SENTENCAS JUDICIAIS RELATIVAS A PESSOAL E ENCARGOS DE INATIVOS",
+  "EXECUCAO ORCAMENTARIA DELEGADA A MUNICIPIOS",
+  "CONSTITUICAO OU AUMENTO DE CAPITAL DE EMPRESAS",
+  "AUXILIO TRANSPORTE",
+  "OUTROS BENEFICIOS ASSISTENCIAIS DO SERVIDOR E DO MILITAR",
+  "INDENIZACOES E RESTITUICOES TRABALHISTAS   INATIVOS",
+  "PRECATORIOS ATIVO CIVIL",
+  "EXECUCAO ORCAMENTARIA DELEGADA A CONSORCIOS PUBLICOS",
+  "TRANSFERENCIAS A CONSORCIOS PUBLICOS MEDIANTE CONTRATO DE RATEIO",
+  "OUTROS BENEFICIOS ASSISTENCIAIS",
+  "AUXILIO FUNERAL",
+  "INDENIZACOES E RESTITUICOES TRABALHISTAS",
+  "OBRIGACOES PATRONAIS",
+  "OUTRAS SENTENCAS JUDICIAIS",
+  "RATEIO PELA PARTICIPACAO EM CONSORCIO PUBLICO",
+  "APLICACAO DIRETA DECORRENTE DE OPERACAO ENTRE ORGAOS  FUNDOS E ENTIDADES INTEGRANTES DOS ORCAMENTOS FISCAL E DA SEGURIDADE SOCIAL",
+  "OUTRAS DESPESAS DE PESSOAL DECORRENTES DE CONTRATOS DE TERCEIRIZACAO",
+  "CONTRIBUICAO PARA ASSISTENCIA SOCIAL DO SERVIDOR",
+  "APORTE PARA COBERTURA DO DEFICIT ATUARIAL DO RPPS",
+  "DESPESAS COM AMORTIZACAO DA DIVIDA REALIZADOS ENTRE ORGAOS  FUNDO E ENTIDADES INTEGRANTES DO ORCAMENTO FISCAL E SEGURIDADE SOCIAL",
+  "OUTRAS INDENIZACOES E RESTITUICOES TRABALHISTAS",
+  "SENTENCAS JUDICIAIS RELATIVAS A PESSOAL E ENCARGOS DE PENSIONISTAS",
+  "COMPENSACOES AO RGPS",
+  "CORR  MONET  CAMBIAL DA DIVIDA CONTRAT  RESGATADA",
+  "CONTRIBUICOES PREVIDENCIARIAS   RPPS   INATIVOS E PENSIONISTAS   PLANO PREVIDENCIARIO",
+  "AUXILIO FINANCEIRO A ESTUDANTES",
+  "ENCARGOS DE PESSOAL REQUISITADO DE OUTROS ENTES",
+  "TRANSFERENCIAS A INSTITUICOES MULTIGOVERNAMENTAIS",
+  "AUXILIO RECLUSAO",
+  "RESERVAS DE CONTINGENCIA",
+  "A CLASSIFICAR",
+  "COR  MONET  DIVIDA OP CREDITO P  ANTECIP  DA RECEITA",
+  "TRANSFERENCIAS A INSTITUICOES PRIVADAS COM FINS LUCRATIVOS",
+  "AUXILIO FINANCEIRO A PESQUISADORES",
+  "TRANSFERENCIAS FUNDO A FUNDO AOS ESTADOS E AO DISTRITO FEDERAL A CONTA DE RECURSOS DE QUE TRATA O ART  25 DA LEI COMPLEMENTAR NO 141  DE 2012 ",
+  "A DEFINIR",
+  "CONTRIBUICOES PREVIDENCIARIAS   RPPS   PESSOAL ATIVO DO PLANO FINANCEIRO",
+  "BENEFICIO MENSAL AO DEFICIENTE E IDOSO",
+  "SEGURO DESEMPREGO E ABONO SALARIAL",
+  "ENCARGOS SOBRE OPERACOES DE CREDITO POR ANTECIPACAO DA RECEITA",
+  "CONVOCACAO EXTRAORDINARIA  INCISO II  6O  ART  57 DA CF ",
+  "ELEMENTO GENERICO   TRANSFERENCIAS A ESTADOS E AO DISTRITO FEDERAL",
+  "PARTICIPACAO EM FUNDOS  ORGANISMOS OU ENTIDADES ASSEMELLHADAS  NACIONAIS E INTERNACIONAIS",
+  "CONTRIBUICOES PREVIDENCIARIAS   RPPS   INATIVOS E PENSIONISTAS   PLANO FINANCEIRO",
+  "AUXILIO NATALIDADE",
+  "DIVIDA CONTRATUAL INTERNA",
+  "CONTRIBUICOES PREVIDENCIARIAS   RPPS   PESSOAL ATIVO DO PLANO PREVIDENCIARIO",
+  "LOCACAO DE MAO DE OBRA",
+  "TRANSFERENCIAS A UNIAO",
+  "AUXILIOS",
+  "TRANSFERENCIAS A MUNICIPIOS",
+  "ELEMENTO GENERICO   TRANSFERENCIAS A ESTADOS E DISTRITO FEDERAL",
+  "TRANSFERENCIAS FUNDO A FUNDO AOS MUNICIPIOS A CONTA DE RECURSOS DE QUE TRATAM OS 1O E 2O DO ART  24 DA LEI COMPLEMENTAR NO 141  DE 2012 ",
+  "AUXILIO SAUDE",
+  "DESP DE EXERCICIOS ANTERIORES RELATIVAS A PESSOAL E ENCARGOS DE INATIVOS E PENSIONISTAS",
+  "JUROS  DESAGIOS E DESCONTOS DA DIVIDA MOBILIARIA",
+  "DIVIDA CONTRATUAL EXTERNA",
+  "CONTRIBUICAO A ENTIDADES FECHADAS DE PREVIDENCIA",
+  "INDENIZACAO PELA EXECUCAO DE TRABALHOS DE CAMPO",
+  "ELEMENTO GENERICO",
+  "AUXILIO FARDAMENTO",
+  "DEMAIS OBRIGACOES PATRONAIS",
+  "INDENIZACOES E RESTITUICOES TRABALHISTAS   PENSIONISTAS",
+  "APLICACAO DIRETA A CONTA DE RECURSOS DE QUE TRATAM OS 1O E 2O DO ART  24 DA LEI COMPLEMENTAR NO 141  DE 2012 ",
+  "TRANSFERENCIAS A CONSORCIOS PUBLICOS MEDIANTE CONTRATO DE RATEIO A CONTA DE RECURSOS DE QUE TRATA O ART  25 DA LEI COMPLEMENTAR NO 141  DE 2012 ",
+  "TRANSFERENCIAS A CONSORCIOS PUBLICOS MEDIANTE CONTRATO DE RATEIO A CONTA DE RECURSOS DE QUE TRATAM OS 1O E 2O DO ART  24 DA LEI COMPLEMENTAR NO 141  DE 2012 ",
+  "ALIQUOTA SUPLEMENTAR DE CONTRIBUICAO PREVIDENCIARIA   PESSOAL ATIVO   PLANO FINANCEIRO",
+  "APLICACAO DIRETA A CONTA DE RECURSOS DE QUE TRATA O ART  25 DA LEI COMPLEMENTAR NO 141  DE 2012 ",
+  "DESP DE EXERCICIOS ANTERIORES RELATIVAS A PESSOAL E ENCARGOS DE PENSIONISTAS",
+  "DESP DE EXERCICIOS ANTERIORES RELATIVAS A CONTRATACAO POR TEMPO DETERMINADO",
+  "OUTROS EMPRESTIMOS E FINANCIAMENTOS",
+  "INCENTIVOS VIA INSTITUICAO FINANCEIRA",
+  "PAGAS COM RECURSOS DO APORTE P COBERTURA DEFICIT ATUARIAL ANTES DE 5 ANOS  RPPS ",
+  "INCENTIVOS DIRETOS A CONTRIBUINTES",
+  "APLICACOES DIRETAS"
+]
